Move mapReceiptOperation to mapping.ts
