+++ src/services/finp2p-contract/escrow.ts
@@ -5,7 +5,8 @@ import {
   logger
 } from "@owneraio/finp2p-nodejs-skeleton-adapter";
 import { ValidationError, EthereumTransactionError } from "@owneraio/finp2p-contracts";
-import { CommonServiceImpl, mapReceiptOperation } from "./common";
+import { CommonServiceImpl } from "./common";
+import { mapReceiptOperation } from "./mapping";
 import { emptyOperationParams, extractBusinessDetails } from "./helpers";
 import { validateRequest } from "./validator";
